refactor(pipeline): reorganize territory logic into domain-focused directory
Move territory sync and validation logic from `stages/` to `territory/` package to reflect domain-driven organization. This aligns the directory structure with the actual business domain boundaries.
