+++ apps/pipeline/src/territory/validate.ts
